backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-jose[cryptography]
httpx
pydantic
pydantic-settings
python-multipart
